talk-demo/app/requirements.txt
Flask==1.1.1
flask-cors==3.0.7
mysql-connector==2.2.9
ptvsd==4.3.2
